wordsmithing ch02.asciidoc #5
In the paragraph beginning "Transactions are like lines in a double-entry bookkeeping ledger."  This paragraph needed to be re-written.  It was incorrect from an accounting interpretation.  Debits (meaning "left") are not really on the left side of the graphic and typical display for Inputs in this case, since this is in effect a balance sheet journal entry.  Feel free to double check this with an actual accountant.