Update ch01.asciidoc, What is bitcoin 3rd para
Made a few implied points more explicit and changed "transaction" to "address" in a few occasions where it was more appropriate
for e.g.  
"Users of bitcoin own keys which allow them to prove ownership of transactions in the bitcoin network, unlocking the value to spend it and transfer it to a new recipient. " to 
"Users of bitcoin own keys which allow them to prove ownership of addresses in the bitcoin network. Unlocking these addresses, unlocks any value associated to it and grants the user the ability to spend and thus transfer value from it."

Do key holders really own transactions? when someone transfers bitcoin to a user's address, is that transaction really theirs rather than the spender's? Isn't it more appropriate to say that the bitcoins associated to their address is theirs.

"Unlocking value to spend it and transfer it to a new recipient"
here unlocking value implies the "value to spend" instead of "spend the value", also, isn't spend and transfer the same thing? and a user could transfer bitcoins to an old recipient instead of a new one.