Update ch02.asciidoc
When I looked at the hard copy of the book, I was concerned about space on the page. The changes I have suggested could have been put closer to Figure 2.8, but as a newbie I was referring to the text, then following the process described in the text as it appeared in Figure 2.8. My vote is that it fits better where I put it. I also tried to use as few words as possible while making sure "Best Practice" was clearly emphasized as early as possible in a newbies investigations/studies/indoctrination. Pick a word.    :)